// frontend/src/app/api/generated/model/guest-balance-account.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GuestBalanceTransaction } from './guest-balance-transaction';


export interface GuestBalanceAccount { 
    guestId: string;
    guestName: string;
    balance: number;
    transactions: Array<GuestBalanceTransaction>;
}